pyproject.toml
[project]
name = "video-demo"
version = "0.1.0"
description = "Video processing demo — FastAPI + Celery + Redis + Postgres + MinIO"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135,<1",
    "uvicorn[standard]>=0.41,<1",
    "pydantic>=2.12,<3",
    "pydantic-settings>=2.13,<3",
    "tortoise-orm[asyncpg]>=1.1,<2",
    "aerich>=0.8,<1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15,<1",
    "mypy>=1.19,<2",
    "pytest>=8,<9",
    "pytest-asyncio>=0.26,<1",
    "httpx>=0.28,<1",
    "testcontainers[postgres]>=4.14,<5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
python-preference = "managed"

[tool.ruff]
target-version = "py314"
line-length = 88

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "SIM",
    "TCH",
    "RUF",
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
